requirements.txt
requests>=2.25.1
pyyaml>=5.4.1 